requirements.txt
d2to1>=0.2.10,<0.3
pbr>=0.5.21,<1.0
eventlet>=0.13.0
Flask>=0.10,<1.0
oslo.config>=1.2.0
PyYAML

#openstack-common requires
iso8601>=0.1.4
kombu>=2.4.8
Babel>=1.3
python-keystoneclient>=0.2.0
